refactor(connectors/pusher): extract options code to util
